Split up reconcile.go
Split up the file to make it easier to maintain.
